Add the action organization initial_import and tests (#992)
* Add draft version of the organization.initial_import action.

Add a test. It is a *draft* version.

* Add an empty datastore check.

Update tests.

* Fix problem with fake/dummy models in registry.

* Mark organization.initial_import action as internal as it should be called from inside the stack.

* Just check organization.

To check if a datastore is empty, check if an organization exists.

* Simplify the code, remove not needed parts.

* Simplify checker and tests.

Add code to clean the registry in the tests, because it is needed in the
tests.

* Check all models in a test automaticly.

* Fix tests.

* Add get_initial_data and use it in the tests.

Remove the bloated simple test. Just rely on the test with the initial
data.

* Update tests.

Some tests inject fake models into the model_registry. Prevent this.
This is needed for organization.initial_import to work.

* Add some more tests for organization.initial_import.

Add four tests to check for data problems. This checks the checker,
which is included in the action.

* Update tests.

* Hardcode the collections in the checker.

This is done like before in the other modes of the checker.
This simplifies the tests *a lot*.