Unify tests for organization scope with permission in meeting
test_two_meetings for set_password action had 3 cases. In all the cases target
user was in 2 meetings (1, 4) in different committees, without
admin rights. Cases 2 and 3 are already tested in other tests for all
actions and could be safely deleted. Case 1 was replaced with a separate test:
1. Test user shares both meetings with the target user but has admin rights
    only in one of them. -> new case for all actions:
    test_scope_organization_permission_in_one_meeting_two_shared_meetings
2. Test user shares one meeting with the target user and has admin rights in it
    -> tested in test_scope_organization_permission_in_one_meeting
3. Test user shares both meetings with the target user and has admin rights in
    them -> tested in test_scope_organization_permission_in_all_meetings

For more clarity renamed test_scope_organization_permission_in_one_meeting
and moved to the helper method the common setup logic used for all 3 cases
and for the case with archieved meetings only.