Fix case 2 (same titles on meeting level)
During meeting-wide mediafile duplication it is now checked if a mediafile with
the same parent_id (or root if None) already exists with a conflicting name:
either the origin title (origin_title) or a variant like 'origin_title (#n)'.
If yes, a new unique title like 'origin_title (#n)' is generated for the
mediafile. If only 'origin_title' is present, it's 'origin_title (#2)',
else n is calculated as max n for this origin_title + 1.